feat: TER-78 Update the job manager API plus few docker setup changes
In this commit, we are updating hte job manager API to expose an endpoint
to vew all jobs available on the jobs docker container. We update the
Vue code to show existing jobs to select from or upload a new job.

We have to add the frontend to make a request to the jobs api directly,
so we update the docker file to add the frontend container to attach to
the jobs docker network, and we also add a proxy route to the route api
endpoint.

We also expose the port 8080 on the frontend docker container.